fix(RAPT-236): When two hotspots point to the same node, with different start time, one option is always randomly chosen (#108)
the issue:
When two hotspots point to the same node, with different start time, one option is always randomly chosen.

root cause:
checkSeekTo function is creating a list of all the hotspots on the current node, then performing a for loop and trying to find a hotspot which its destination id matches the next node id and that has a startFrom. these validations are not enough since more than 1 hotspot can fulfill them. Hence, a different hotspot from the one that was clicked can be chosen.

solution:
when invoking checkSeekTo function, we know the hotspotId that was clicked; solution is to pass the id to the function and extract the startFrom value from the hotspot data.

Solves RAPT-236