fix(SUP-49327): [Novartis] Multiple rapt media embedded in single page issue (#146)
Issue:
When embedding two rapt player in one single page clicking on hotspots on second player cause the hotspots on the first player to disappear.

Root Cause:
we are selecting the kiv-rapt-engine element directly from the dom so we always get the element from first player
(first element in the dom with this class) and not from relevant player.

Solution:
Select the kiv-rapt-engine from the current rapt player